\name{qat_data_close_ncdf}
\alias{qat_data_close_ncdf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Close an open netCDF-file}
\description{
An open netCDF file will be closed.
}
\usage{
qat_data_close_ncdf(obj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{An open netCDF object.}
}
\value{
None.
}
\author{Andre Duesterhus}
\seealso{\code{\link{qat_data_read_ncdf}}, \code{\link{qat_data_nameofvars_ncdf}}, \code{\link{qat_data_numofvars_ncdf}}, \code{\link{qat_data_varcontent_ncdf}}}
\examples{
#still to come
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{IO}